Run morph segmentation after updating orthography
If the segmentation runs on old orthography, the morphs won't
correspond to the lemma. But simply switching the two modules doesn't
work either, because the orthography update forcibly rewrites lemmas,
messing up internal state. The solution is to save and reload the file
in between, ensuring internal consistency of the Lexicon object.